added a button to start the mission
Since I keep getting an error when starting the mission, instead of starting it in the completion block, i create a new button that is revealed in the completion block. The user is then prompted to push it. I also added a button to remove the waypoint from the map and the mission.